[cli] Check if node is synced (#942)
This adds a new cli command to check if the node is synced. It also includes a check in the base command to make sure the node is synced before proceeding with commands by default. For a few commands that don't require the node to be synced, a configuration is included to skip that requirement.


* add command to check if synced

* put the command in the right place

* add a helper function to check if synced, handling the case w=in which syncing has not (re)started, and the latest block is old

* Add a helper function to require a node is synced. Use this function at the beginning of any commands that need the node to be synced

* remove superfluous console.log

* regenerate docs to include the new command

* check if synced in base command. add ability to skip that check

* account:new doesn't need the node synced