Correct documentation on the validator and validator group deregister (#5197)
### Description

The documentation on the `celocli validator:deregister` and `celocli validatorgroup:deregister` describe an outdated version of the Valdiator contract code. It's current implementation requires validators to leave their groups before attempting to deregister, whereas the waiting period was attached to withdrawing locked gold before.

This PR updates the doc string on these commands to correct the descriptions.

https://discordapp.com/channels/600834479145353243/601046166620078100/758486506075324437

_Brief explanation of why these changes are/are not backwards compatible._